requirements.txt
sqlalchemy~=2.0.31
bcrypt~=4.1.3
python-dotenv~=1.0.1
streamlit~=1.36.0
langchain~=0.2.6
pandas~=2.2.2
fuzzywuzzy~=0.18.0
matplotlib~=3.9.1
numpy~=1.26.4